fruit-defect/requirements.txt
torch
torchvision
torchaudio
pillow
numpy
pyyaml
tqdm
scikit-learn
tensorboard
minio